Redid db schema a bit and fixed some issues
* the share model was unwieldy, changed to using  permissions table
* fixed some overlap with built in list class

+++ app/main/routes.py
@@ -1,7 +1,7 @@
 from datetime import date
 from flask import render_template, redirect, url_for, jsonify, request, flash
 from flask_login import current_user, login_required
-from app.models import List, Entry, User, Share
+from app.models import List, Entry, User, ListPermission
 from app import db
 from app.main import bp
 from app.main.decorators import check_list_access, check_entry_access, check_list_owner
@@ -35,7 +35,7 @@ def delete_list(list_id):
       return redirect(url_for('main.index'))
     else:
       return redirect(url_for('main.index'))
-  return render_template('confirm_delete.html', form=form, list=list_)
+  return render_template('confirm_delete.html', form=form, list_=list_)
 
 
 @bp.route('/get_api_key/<list_id>')
@@ -43,13 +43,14 @@ def delete_list(list_id):
 @check_list_access
 def get_api_key(list_id):
   list_ = List.query.filter_by(id=list_id).first()
-  return render_template('get_api_key.html', list=list_)
+  return render_template('get_api_key.html', list_=list_)
 
 
 @bp.route('/lists')
 @login_required
 def get_lists():
-  return render_template('lists.html')
+  lists = current_user.get_lists()
+  return render_template('lists.html', lists=lists)
 
 
 @bp.route('/list/<list_id>')
@@ -63,7 +64,7 @@ def get_list(list_id):
       d,
       d + sett.days_to_display
   )
-  return render_template('list.html', list=list_, days=days)
+  return render_template('list.html', list_=list_, days=days)
 
 
 @bp.route('/new_list', methods=['GET', 'POST'])
@@ -71,9 +72,12 @@ def get_list(list_id):
 def new_list():
   form = NewListForm()
   if form.validate_on_submit():
-    list_ = List(name=form.name.data, owner_id=current_user.id)
+    list_ = List(name=form.name.data)
     list_.generate_api_key()
     db.session.add(list_)
+    db.session.commit()
+    perm = ListPermission(list_id=list_.id, user_id=current_user.id, permission_level='owner')
+    db.session.add(perm)
     db.session.commit()
     return redirect(url_for('main.get_list', list_id=list_.id))
   return render_template('new_list.html', form=form)
@@ -105,7 +109,7 @@ def list_settings(list_id):
     if curr_settings.days_to_display != form.daysToDisplay.data:
       curr_settings.days_to_display = form.daysToDisplay.data
     db.session.commit()
-  return render_template('list_settings.html', list=list_, form=form)
+  return render_template('list_settings.html', list_=list_, form=form)
 
 
 @bp.route('/list_foods/<list_id>')
@@ -114,7 +118,7 @@ def list_settings(list_id):
 def list_foods(list_id):
   # TODO Create a list of meals that can be autocompleted
   list_ = List.query.filter_by(id=list_id).first_or_404()
-  return render_template('list_foods.html', list=list_)
+  return render_template('list_foods.html', list_=list_)
 
 
 @bp.route('/list_daily_meals/<list_id>')
@@ -123,16 +127,17 @@ def list_foods(list_id):
 def list_daily_meals(list_id):
   # TODO List and change the different meals of the day i.e. lunch, dinner, breakfast etc.
   list_ = List.query.filter_by(id=list_id).first_or_404()
-  return render_template('list_daily_meals.html', list=list_)
+  return render_template('list_daily_meals.html', list_=list_)
 
 
-@bp.route('/drop_share/<list_id>/<share_id>')
+@bp.route('/drop_share/<list_id>/<user_id>')
 @login_required
 @check_list_access
-def drop_share(list_id, share_id):
-  s = Share.query.filter_by(id=share_id, list_id=list_id).first()
-  db.session.delete(s)
-  db.session.commit()
+def drop_share(list_id, user_id):
+  s = ListPermission.query.filter_by(user_id=user_id, list_id=list_id).first()
+  if s.permission_level != 'owner':
+    db.session.delete(s)
+    db.session.commit()
   return redirect(url_for('main.list_shares', list_id=list_id))
 
 
@@ -142,21 +147,21 @@ def drop_share(list_id, share_id):
 def list_shares(list_id):
   form = ShareForm()
   list_ = List.query.filter_by(id=list_id).first_or_404()
-  l_shares = list_.shares
+  l_shares = list_.users
   if form.validate_on_submit():
     u = User.query.filter(
         (User.username == form.target.data.lower()) | (User.email == form.target.data.lower())
     ).first()
-    s = Share.query.filter_by(
-        list_id=list_.id, grantee_id=u.id
+    s = ListPermission.query.filter_by(
+        list_id=list_.id, user_id=u.id
     ).first()
     if not s:
-      s = Share(list_id=list_.id, owner_id=list_.owner_id, grantee_id=u.id)
+      s = ListPermission(list_id=list_.id, user_id=u.id, permission_level='rw')
       db.session.add(s)
       db.session.commit()
     return redirect(url_for('main.list_shares', list_id=list_.id))
   return render_template('list_shares.html',
-                         list=list_,
+                         list_=list_,
                          form=form,
                          list_shares=l_shares
                         )
